refactor: split scripts toolset into business-rules, client-scripts, script-includes
Replaces the coarse `scripts` domain with three per-artifact toolsets so
gateway sessions can scope the tool surface precisely. Each new folder
(business-rules/, client-scripts/, script-includes/) is a self-contained
domain with its own index.ts, schemas.ts, tool file, and colocated tests.
TOOLSETS in registry.ts and the X-MCP-Toolsets docs in README are updated
to reflect the three new names.
