+++ src/tools/client-scripts/client-scripts.test.ts
