+++ src/tools/client-scripts/client-scripts.ts
